Set up redux-devtools
